Add files via upload
University of Limerick students are tasked with a Final Year Project at the end of their undergraduate degree. This project is based on a topic of my own choosing, and was the product of wanting to satisfy my own curiosity about how mathematical modelling techniques could be used to calculate the force acting on a wind turbine.

The project brings together fluid dynamics, advanced complex number concepts. Extensive MATLAB programming was used to produce diagrams and visuals.